Return value of the function 'hash()' set to 'unsigned long' following the body of the function.  Previous return value of 'unsigned' defaults to 'unsigned int'.

+++ src/sym_table.c
@@ -121,7 +121,7 @@ register ub1 *k;        /* the key */
 /* ================================================================ */
 
 
-unsigned hash(char *sym)
+unsigned long hash(char *sym)
 {
   ub4 hashval;
